+++ node/src/errors/http.ts
@@ -15,10 +15,22 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+import * as axios from "axios";
+
 export class HTTPError extends Error {
     public readonly status: number;
     constructor(
         public readonly response: Response,
+    ) {
+        super(response.statusText);
+        this.status = response.status;
+    }
+}
+
+export class AxiosHTTPError extends Error {
+    public readonly status: number;
+    constructor(
+        public readonly response: axios.AxiosResponse,
     ) {
         super(response.statusText);
         this.status = response.status;