+++ node/src/types/kernel/api/notification/pushMsg/response.d.ts
@@ -23,7 +23,7 @@ export interface IResponse {
      * status code
      */
     code: number;
-    data: Data;
+    data: IData;
     /**
      * status message
      */
@@ -31,9 +31,9 @@ export interface IResponse {
 }
 
 /**
- * push alert information
+ * the related info of the notification
  */
-export interface Data {
+export interface IData {
     /**
      * message ID
      */