+++ node/src/types/kernel/index.d.ts
@@ -1,5 +1,10 @@
 export type ID = string; // ID 类型
 
+/* 请求体 */
+export interface IPayload extends object {
+    
+}
+
 /* 响应体 */
 export interface IResponse {
     code: number;